requirements.txt
absl-py==0.8.1
amqp==2.6.1
appdirs==1.4.4
astor==0.8.0
attrs==19.3.0
base==1.0.4
Bashutils==0.0.4
billiard==3.6.3.0
bleach==3.1.5
cachetools==3.1.1
celery==4.4.7
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cmake==3.15.3
colorama==0.4.3
cycler==0.10.0
data==0.4
decorator==4.4.1
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.1
dlib==19.8.1
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-Script==2.0.6
funcsigs==1.0.2
gast==0.2.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.19.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.52.0
grpcio==1.25.0
gspread==3.1.0
h5py==2.10.0
httplib2==0.15.0
idna==2.8
imageio==2.6.1
importlib-metadata==1.7.0
importlib-resources==3.0.0
imutils==0.5.3
ipykernel==5.3.4
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.1.0
kombu==4.6.11
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
multiprocess==0.70.9
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nms==0.1.6
notebook==6.0.3
numpy==1.17.3
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
packaging==20.4
pandas==0.25.2
pandocfilters==1.4.2
pathos==0.2.5
pickleshare==0.7.5
Pillow==6.2.0
pox==0.2.7
ppft==1.6.6.1
prometheus-client==0.8.0
prompt-toolkit==1.0.18
props==0.0.3
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.6.1
pyparsing==2.4.2
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.1.2
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
redis==3.5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.0
torchvision==0.4.1
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.7
vine==1.3.0
virtualenv==20.0.26
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
win-unicode-console==0.5
wrapt==1.11.2
zipp==3.1.0
